Create higher level PVB API wrapper
This handles the request paths and response parsing
